Update a bunch of versions, and publish to GH pkgs on main
I've added a new flow to push packages to github packages when things
are merged to main. This will enable testing before a release to maven
central and the gradle plugins portal.

I've also added a bunch of libraries that I like using, so that I can
use them downstream in the examples project.